src/requirements.txt
torch
torchvision
transformers
albumentations
numpy
pillow
tqdm
tensorboard
